Refactored to avoid rounding error using BigDecimal

+++ src/main/java/tdd/vendingMachine/Product.java
@@ -1,13 +1,15 @@
 package tdd.vendingMachine;
 
+import java.math.BigDecimal;
+
 /**
  * Created by sok_pomaranczowy on 19.03.17.
  */
 public class Product {
     private String name;
-    private Double price;
+    private BigDecimal price;
 
-    public Product(String name, Double price) {
+    public Product(String name, BigDecimal price) {
         this.name = name;
         this.price  = price;
     }
@@ -16,7 +18,7 @@ public class Product {
         return name;
     }
 
-    public Double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
